+++ bauh/api/paths.py
@@ -1,8 +1,9 @@
 from pathlib import Path
 
+from bauh import __app_name__
 from bauh.api import user
 
-CACHE_PATH = '{}/.cache/bauh'.format(str(Path.home()))
+CACHE_PATH = f'/var/cache/{__app_name__}' if user.is_root() else f'{Path.home()}/.cache/{__app_name__}'
 CONFIG_PATH = '{}/.config/bauh'.format(str(Path.home()))
 USER_THEMES_PATH = '{}/.local/share/bauh/themes'.format(str(Path.home()))
 DESKTOP_ENTRIES_DIR = '{}/.local/share/applications'.format(str(Path.home()))