+++ bauh/gems/arch/database.py
@@ -10,7 +10,7 @@ from typing import Optional
 from bauh.api.paths import CACHE_PATH
 from bauh.commons.system import ProcessHandler
 
-SYNC_FILE = '{}/arch/db_sync'.format(CACHE_PATH)
+SYNC_FILE = f'{CACHE_PATH}/arch/db_sync'
 
 
 def should_sync(arch_config: dict, aur_supported: bool, handler: Optional[ProcessHandler], logger: logging.Logger):