+++ bauh/gems/snap/__init__.py
@@ -1,12 +1,12 @@
 import os
 
-from bauh.api.paths import CACHE_PATH, CONFIG_PATH
+from bauh.api.paths import CONFIG_PATH, CACHE_PATH
 from bauh.commons import resource
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-SNAP_CACHE_PATH = CACHE_PATH + '/snap'
+SNAP_CACHE_PATH = f'{CACHE_PATH}/snap'
 CONFIG_FILE = '{}/snap.yml'.format(CONFIG_PATH)
-CATEGORIES_FILE_PATH = SNAP_CACHE_PATH + '/categories.txt'
+CATEGORIES_FILE_PATH = f'{SNAP_CACHE_PATH}/categories.txt'
 URL_CATEGORIES_FILE = 'https://raw.githubusercontent.com/vinifmor/bauh-files/master/snap/categories.txt'
 SUGGESTIONS_FILE = 'https://raw.githubusercontent.com/vinifmor/bauh-files/master/snap/suggestions.txt'